Refactor user and notification handling; update service methods, enhance error handling, and improve DTOs for notifications

+++ common/src/main/java/org/example/service/impl/UserServiceImpl.java
@@ -3,7 +3,7 @@ package org.example.service.impl;
 import jakarta.mail.MessagingException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.example.dto.notification.NotificationResponseDto;
+import org.example.dto.notification.NotificationRequestDto;
 import org.example.dto.user.ChangePasswordRequest;
 import org.example.dto.user.ResetPasswordRequest;
 import org.example.dto.user.UserRegisterDto;
@@ -22,6 +22,9 @@ import org.example.service.NotificationService;
 import org.example.service.SendMailService;
 import org.example.service.UserService;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -29,7 +32,6 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
-import java.util.Optional;
 import java.util.Random;
 
 
@@ -51,47 +53,61 @@ public class UserServiceImpl implements UserService {
     private final Random random = new Random();
 
     @Override
-    public Optional<UserRegisterDto> findByEmail(String username) {
-        return userRepository.findByEmail(username).map(userRegisterMapper::toUserRegisterDto);
+    public UserRegisterDto findByEmail(String username) {
+        return userRepository.findByEmail(username)
+                .map(userRegisterMapper::toUserRegisterDto)
+                .orElse(null);
     }
 
     @Override
-    public Optional<ChangePasswordRequest> changePassword(String email, String oldPassword, String newPassword) {
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND_BY_EMAIL, email));
+    public ChangePasswordRequest changePassword(String email,
+                                                String oldPassword,
+                                                String newPassword) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new BusinessException(
+                        ErrorCode.USER_NOT_FOUND_BY_EMAIL, email));
         if (!passwordEncoder.matches(oldPassword, user.getPassword())) {
             throw new BusinessException(ErrorCode.OLD_PASSWORD_IS_INCORRECT);
         }
         user.setPassword(passwordEncoder.encode(newPassword));
         User savedUser = userRepository.save(user);
 
-        NotificationResponseDto notificationResponseDto = new NotificationResponseDto();
-        notificationResponseDto.setUserId(savedUser.getId());
-        notificationResponseDto.setTitle("You changed your password successfully!");
-        notificationResponseDto.setMessage(NotificationType.PROFILE_PASSWORD_CHANGED_NOTIFICATION.format(savedUser.getName(), savedUser.getSurname()));
-        notificationService.save(notificationResponseDto);
+        NotificationRequestDto notificationRequestDto = new NotificationRequestDto();
+        notificationRequestDto.setUserId(savedUser.getId());
+        notificationRequestDto.setTitle("You changed your password successfully!");
+        notificationRequestDto.setMessage(
+                NotificationType.PROFILE_PASSWORD_CHANGED_NOTIFICATION
+                        .format(savedUser.getName(), savedUser.getSurname()));
+        notificationService.save(notificationRequestDto);
 
-        return Optional.of(changePasswordRequestMapper.toChangePasswordRequestDto(savedUser));
+        return changePasswordRequestMapper.toChangePasswordRequestDto(savedUser);
     }
 
     @Override
-    public Optional<ChangePasswordRequest> changePasswordByEmail(String email) {
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND_BY_EMAIL, email));
+    public ChangePasswordRequest changePasswordByEmail(String email) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND_BY_EMAIL,
+                        email));
         String verificationCode = generateVerificationCode();
         try {
             sendMailService.sendVerificationMailHtml(user.getEmail(), verificationCode);
             user.setVerificationCode(verificationCode);
             userRepository.save(user);
-            return Optional.of(changePasswordRequestMapper.toChangePasswordRequestDto(user));
+            return changePasswordRequestMapper.toChangePasswordRequestDto(user);
         } catch (MessagingException e) {
             e.printStackTrace();
-            return Optional.empty();
+            throw new BusinessException(ErrorCode.TRY_AGAIN);
         }
     }
 
     @Override
-    public Optional<ResetPasswordRequest> resetPassword(String email, String code, String newPassword, String newConfirmPassword) {
+    public ResetPasswordRequest resetPassword(String email,
+                                              String code,
+                                              String newPassword,
+                                              String newConfirmPassword) {
         User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND_BY_EMAIL, email));
+                .orElseThrow(() -> new BusinessException(
+                        ErrorCode.USER_NOT_FOUND_BY_EMAIL, email));
 
         if (!verifyUser(email, code)) {
             throw new BusinessException(ErrorCode.VERIFICATION_FAILED, email);
@@ -104,29 +120,25 @@ public class UserServiceImpl implements UserService {
         user.setPassword(passwordEncoder.encode(newPassword));
         User savedUser = userRepository.save(user);
 
-        NotificationResponseDto notificationResponseDto = new NotificationResponseDto();
-        notificationResponseDto.setUserId(savedUser.getId());
-        notificationResponseDto.setTitle("You reset your password successfully!");
-        notificationResponseDto.setMessage(NotificationType.PROFILE_PASSWORD_RESET_NOTIFICATION.format(savedUser.getName(), savedUser.getSurname()));
-        notificationService.save(notificationResponseDto);
+        NotificationRequestDto notificationRequestDto = new NotificationRequestDto();
+        notificationRequestDto.setUserId(savedUser.getId());
+        notificationRequestDto.setTitle("You reset your password successfully!");
+        notificationRequestDto.setMessage(NotificationType
+                .PROFILE_PASSWORD_RESET_NOTIFICATION
+                .format(savedUser.getName(), savedUser.getSurname()));
+        notificationService.save(notificationRequestDto);
 
-        return Optional.of(userResetPasswordMapper.toUserResetPasswordDto(savedUser));
+        return userResetPasswordMapper.toUserResetPasswordDto(savedUser);
     }
 
     @Override
-    public UserRegisterDto save(UserRegisterDto userRegisterDto, MultipartFile multipartFile) {
-        if (multipartFile != null && !multipartFile.isEmpty()) {
-            String fileName = System.currentTimeMillis() + "_" + multipartFile.getOriginalFilename();
-            File file = new File(imageDirectoryPath + fileName);
-            try {
-                multipartFile.transferTo(file);
-                userRegisterDto.setPicName(fileName);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        } else {
-            userRepository.findById(userRegisterDto.getId())
-                    .ifPresent(productOptional -> userRegisterDto.setPicName(productOptional.getPicName()));
+    public UserRegisterDto save(UserRegisterDto userRegisterDto,
+                                MultipartFile multipartFile) {
+        userAndMultipartFile(userRegisterDto, multipartFile);
+        if (userRegisterDto.getId() != 0) {
+            userRepository.findById(userRegisterDto.getId()).ifPresent(
+                    productOptional ->
+                            userRegisterDto.setPicName(productOptional.getPicName()));
         }
         User user = userRegisterMapper.toUser(userRegisterDto);
         user.setPassword(passwordEncoder.encode(user.getPassword()));
@@ -139,19 +151,26 @@ public class UserServiceImpl implements UserService {
         }
         User savedUser = userRepository.save(user);
 
-        NotificationResponseDto notificationResponseDto = new NotificationResponseDto();
-        notificationResponseDto.setUserId(savedUser.getId());
-        notificationResponseDto.setTitle("You registered successfully!");
-        notificationResponseDto.setMessage(NotificationType.USER_REGISTERED_NOTIFICATION.format(savedUser.getName(), savedUser.getSurname()));
-        notificationService.save(notificationResponseDto);
+        NotificationRequestDto notificationRequestDto = new NotificationRequestDto();
+        notificationRequestDto.setUserId(savedUser.getId());
+        notificationRequestDto.setTitle("You registered successfully!");
+        notificationRequestDto.setMessage(NotificationType.USER_REGISTERED_NOTIFICATION
+                .format(savedUser.getName(), savedUser.getSurname()));
+        notificationService.save(notificationRequestDto);
 
         return userRegisterMapper.toUserRegisterDto(savedUser);
     }
 
+    @Override
+    public UserRegisterDto createManager(UserRegisterDto userRegisterDto,
+                                         MultipartFile file) {
+        userRegisterDto.setRole(Role.MANAGER);
+        return save(userRegisterDto, file);
+    }
 
     @Override
     public List<UserRequestDto> findAll() {
-        return userRepository.findAll().stream().map(userRequestMapper::toUserRequestDto).toList();
+        return userRequestMapper.toUserRequestDtoList(userRepository.findAll().stream().filter(user -> user.getRole() != Role.ADMIN).toList());
     }
 
 
@@ -165,6 +184,9 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deleteById(int id) {
         userRepository.findById(id).ifPresent(user -> {
+            if (!chekUserById(id)) {
+                throw new BusinessException(ErrorCode.CANNOT_DELETE_OWN_ACCOUNT);
+            }
             userRepository.deleteById(user.getId());
         });
     }
@@ -172,7 +194,6 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void toggleUserBlockStatus(int id) {
-                userRepository.findById(id);
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND, id));
 
@@ -189,58 +210,89 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Optional<UserRegisterDto> findById(int id) {
-        return userRepository.findById(id).map(userRegisterMapper::toUserRegisterDto);
+    public UserRequestDto findById(int id) {
+        return userRequestMapper.toUserRequestDto(userRepository.findById(id)
+                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND, id)));
     }
 
-
     @Override
-    public UserRegisterDto update(UserRegisterDto userRegisterDto, MultipartFile multipartFile) {
-        User existingUser = userRepository.findById(userRegisterDto.getId())
-                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND, userRegisterDto.getId()));
-
-        existingUser.setName(userRegisterDto.getName());
-        existingUser.setSurname(userRegisterDto.getSurname());
-        existingUser.setEmail(userRegisterDto.getEmail());
-        existingUser.setPhone(userRegisterDto.getPhone());
-        existingUser.setPassportDetails(userRegisterDto.getPassportDetails());
-        if (multipartFile != null && !multipartFile.isEmpty()) {
-            String fileName = System.currentTimeMillis() + "_" + multipartFile.getOriginalFilename();
-            File file = new File(imageDirectoryPath + fileName);
-            try {
-                multipartFile.transferTo(file);
-                existingUser.setPicName(fileName);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
+    public UserRegisterDto update(UserRegisterDto userRegisterDto,
+                                  MultipartFile multipartFile) {
+        User user = userRepository.findById(userRegisterDto.getId())
+                .orElseThrow(() -> new BusinessException(
+                        ErrorCode.USER_NOT_FOUND, userRegisterDto.getId()));
+        if (!chekUserById(user.getId())) {
+            throw new BusinessException(ErrorCode.CANNOT_UPDATE_OWN_ACCOUNT);
         }
-        User savedUser = userRepository.save(existingUser);
+        UserRegisterDto userDto = userRegisterMapper.toUserRegisterDto(user);
+        userDto.setName(userRegisterDto.getName());
+        userDto.setSurname(userRegisterDto.getSurname());
+        userDto.setEmail(userRegisterDto.getEmail());
+        userDto.setPhone(userRegisterDto.getPhone());
+        userDto.setPassportDetails(userRegisterDto.getPassportDetails());
+        userAndMultipartFile(userDto, multipartFile);
+        User savedUser = userRepository.save(userRegisterMapper.toUser(userDto));
 
-        NotificationResponseDto notificationResponseDto = new NotificationResponseDto();
-        notificationResponseDto.setUserId(savedUser.getId());
-        notificationResponseDto.setTitle("You update your profile successfully!");
-        notificationResponseDto.setMessage(NotificationType.PROFILE_UPDATE_NOTIFICATION.format(savedUser.getName(), savedUser.getSurname()));
-        notificationService.save(notificationResponseDto);
+        NotificationRequestDto notificationRequestDto = new NotificationRequestDto();
+        notificationRequestDto.setUserId(savedUser.getId());
+        notificationRequestDto.setTitle("You update your profile successfully!");
+        notificationRequestDto.setMessage(NotificationType.
+                PROFILE_UPDATE_NOTIFICATION.
+                format(savedUser.getName(), savedUser.getSurname()));
+        notificationService.save(notificationRequestDto);
 
         return userRegisterMapper.toUserRegisterDto(savedUser);
     }
 
     @Override
     public List<UserRequestDto> findAllByRoleIn(List<Role> roles) {
-        return userRepository.findAllByRoleIn(roles).stream().map(userRequestMapper::toUserRequestDto).toList();
+        return userRequestMapper.toUserRequestDtoList(userRepository.findAllByRoleIn(roles));
     }
 
     @Override
     public boolean verifyUser(String email, String verifyCode) {
-        Optional<User> user = userRepository.findByEmail(email);
-        if (user.isPresent() && user.get().getVerificationCode() != null
-                && user.get().getVerificationCode().equals(verifyCode)) {
-            user.get().setVerificationCode(null);
-            user.get().setBlocked(false);
-            userRepository.save(user.get());
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND, email));
+        if (user.getVerificationCode() != null && user.getVerificationCode().equals(verifyCode)) {
+            user.setVerificationCode(null);
+            user.setBlocked(false);
+            userRepository.save(user);
             return true;
         }
         return false;
+    }
+
+    @Override
+    public boolean chekUserById(int id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+        if (auth == null || !auth.isAuthenticated() || auth instanceof AnonymousAuthenticationToken) {
+            throw new BusinessException(ErrorCode.USER_NOT_AUTHENTICATED);
+        }
+
+        String email = auth.getName();
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new BusinessException(
+                        ErrorCode.USER_NOT_FOUND_BY_EMAIL, email));
+        if (user.getId() != id) {
+            throw new BusinessException(ErrorCode.USER_NOT_FOUND, id);
+        }
+        return true;
+    }
+
+    @Override
+    public void removeUserPicture(UserRequestDto userRequestDto) {
+        User user = userRepository.findById(userRequestDto.getId())
+                .orElseThrow(() -> new BusinessException(
+                        ErrorCode.USER_NOT_FOUND, userRequestDto.getId()));
+        if (user.getPicName() != null) {
+            File file = new File(imageDirectoryPath + user.getPicName());
+            if (file.exists()) {
+                file.delete();
+            }
+            user.setPicName(null);
+            userRepository.save(user);
+        }
     }
 
     private String generateVerificationCode() {
@@ -248,5 +300,18 @@ public class UserServiceImpl implements UserService {
         return String.valueOf(code);
     }
 
-
+    private void userAndMultipartFile(UserRegisterDto userRegisterDto,
+                                      MultipartFile multipartFile) {
+        if (multipartFile != null && !multipartFile.isEmpty()) {
+            String fileName = System.currentTimeMillis() + "_" +
+                    multipartFile.getOriginalFilename();
+            File file = new File(imageDirectoryPath + fileName);
+            try {
+                multipartFile.transferTo(file);
+                userRegisterDto.setPicName(fileName);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+    }
 }
